more fixes suggested by findbugs
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@6206 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/html/HtmlArea.java
@@ -239,7 +239,7 @@ public class HtmlArea extends HtmlElement {
             catch (final NumberFormatException nfe) {
                 throw new NumberFormatException("Circle radius of " + radiusString + " is not yet implemented.");
             }
-            final Ellipse2D ellipse = new Ellipse2D.Double(centerX - radius / 2, centerY - radius / 2,
+            final Ellipse2D ellipse = new Ellipse2D.Double(centerX - (double) radius / 2, centerY - (double) radius / 2,
                     radius, radius);
             if (ellipse.contains(x, y)) {
                 return true;
